feat(youtube,cron): implement youtube channel watchlist automation
This new feature allows authenticated API users to submit valid YouTube channel IDs
or URLs in a POST request to a particular endpoint, which writes that channel to
a saved watchlist.  When a channel is added to the watchlist, it is checked for
YouTube videos, depending on optional "length" and "publishedAfter" parameters, and
the YouTube videos are added as submissions to the ESOVDB.  After the videos are
added, the channel entry in the watchlist is updated with the current date and moved
to the end of the list.

A cron job has also been set up to be scheduled when the server starts, which runs
every hour on the hour and takes the first channel in the watchlist and performs
a YouTube request for that channel using its "lastChecked" property for the
"publishedAfter" parameter in the YouTube API, and any videos since it was last
checked are added to the ESOVDB. Since the channel is moved to the end of the watch-
list after every run, the first item in the watchlist is always checked, and this way
every watched channel gets a chance to be checked.